% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfpred.r
\name{tfpred}
\alias{tfpred}
\title{Predicts transcription factors driving differential gene expression based on a list of
differentially expressed genes}
\usage{
tfpred(gene_ids, species = "hsa")
}
\arguments{
\item{gene_ids}{vector of Entrez Gene IDs of differentially expressed genes}

\item{species}{three-letter species ID (currently, function works only with human genes ('hsa'))}
}
\value{
data frame with transcription factors ranked according to their probability scores
}
\description{
Predicts transcription factors driving differential gene expression based on a list of
differentially expressed genes
}
\examples{
tfpred(c(5714, 135114, 26958, 5899, 494514, 26502, 388325, 79007, 4688, 1428, 259307))

}
